Backend/requirements.txt
flask
flask_cors
textblob
vaderSentiment
gtts
deep_translator
yake
google.generativeai